+++ src/flows/Termination/TerminationFlow.tsx
@@ -1,9 +1,7 @@
 import { TerminationContext } from '@/src/flows/Termination/context';
 import React, { PropsWithChildren, useId } from 'react';
-import { useForm } from 'react-hook-form';
 import { useTermination } from '@/src/flows/Termination/hooks';
 import { JSFModify } from '@/src/flows/CostCalculator/types';
-import { useJsonSchemasValidationFormResolver } from '@/src/components/form/yupValidationResolver';
 
 function TerminationFlowProvider({
   render,
@@ -16,23 +14,11 @@ function TerminationFlowProvider({
   }) => React.ReactNode;
   terminationBag: ReturnType<typeof useTermination>;
 }>) {
-  const resolver = useJsonSchemasValidationFormResolver(
-    // @ts-expect-error no matching type
-    terminationBag.handleValidation,
-  );
-
   const formId = useId();
-  const form = useForm({
-    resolver,
-    defaultValues: terminationBag.initialValues,
-    shouldUnregister: true,
-    mode: 'onBlur',
-  });
 
   return (
     <TerminationContext.Provider
       value={{
-        form,
         formId: formId,
         terminationBag,
       }}
